refactor to lower code duplication
