+++ libraries/tests/e2e/functions/TestUtils/FunctionConstructProps.cs
@@ -1,0 +1,13 @@
+using Amazon.CDK.AWS.Lambda;
+
+namespace TestUtils;
+
+public class FunctionConstructProps
+{
+    public Architecture Architecture;
+    public Runtime Runtime;
+    public string Name;
+    public string Handler;
+    public string SourcePath;
+    public string DistPath;
+}
