+++ libraries/tests/e2e/functions/core/logging/AOT-Function/src/AOT-Function/AOT-Function.csproj
@@ -25,4 +25,9 @@
     <ItemGroup>
       <ProjectReference Include="..\..\..\..\..\..\..\..\src\AWS.Lambda.Powertools.Logging\AWS.Lambda.Powertools.Logging.csproj" />
     </ItemGroup>
+    <ItemGroup>
+      <Compile Include="..\..\..\Function\src\Function\TestHelper.cs">
+        <Link>TestHelper.cs</Link>
+      </Compile>
+    </ItemGroup>
 </Project>
